functions/scheduler/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# azure-storage-blob~=12.10.0
psycopg2-binary~=2.9.1
numpy~=1.21.2
pandas~=1.3.2
SQLAlchemy~=1.4.25
cvxopt==1.2.5
cvxpy==1.1.13
scipy==1.7.1